Pouvoir administrer les ré-inscriptions #199

+++ src/Entity/SecondHand.php
@@ -88,12 +88,12 @@ class SecondHand
         return $this;
     }
 
-    public function getPrice(): float
+    public function getPrice(): int
     {
         return $this->price;
     }
 
-    public function setPrice(float $price): static
+    public function setPrice(int $price): static
     {
         $this->price = $price;
